Move dummy data to separate file and add some todos
